Se reparo un error en el logger y en el manejador de errores

+++ src/controllers/product.controller.js
@@ -1,5 +1,9 @@
 import { productRepository } from "../repositories/repository.js";
 import { productSeeder } from "../seeder/productSeeder.js";
+import handleError from "../services/errors/errorHandler.service.js";
+import EErrors from "../services/errors/errorHandler.dictionary.js";
+import { generateProductError } from "../services/errors/messages/errorMessages.js";
+import logger from "../middlewares/logger/logger.middleware.js";
 
 export const getProductsController = async (request, response) => {
   try {
@@ -29,18 +33,42 @@ export const getProductsByIdController = async (request, response) => {
  */
 export const createProductController = async (request, response) => {
   let product = request.body;
+  if (
+    !product.title ||
+    !product.description ||
+    !product.price ||
+    !product.thumbnail ||
+    !product.stock ||
+    !product.code
+  ) {
+    return handleError.createError({
+      name: "Product creation error",
+      cause: generateProductError(product),
+      message: "Error trying to create the product",
+      code: EErrors.INVALID_PRODUCT_ERROR,
+    });
+  }
   try {
     const getProducts = await productRepository.getProducts(request);
     const result = await productRepository.createProduct(product);
     request.app.get("socketio").emit("updateProducts", getProducts.payload);
     return response.sendSuccess(result);
   } catch (error) {
+    logger.fatal("Error trying to create a product");
     return response.sendServerError(error.message);
   }
 };
 
 export const deleteProductController = async (request, response) => {
   let id = request.params.pid;
+  if (!id) {
+    return handleError.createError({
+      name: "Product deletion error",
+      cause: generateProductError(product),
+      message: "Error trying to delete the product",
+      code: EErrors.INVALID_PRODUCT_ERROR,
+    });
+  }
   try {
     const result = await productRepository.deleteProduct(id);
     if (result.error) {
@@ -55,12 +83,25 @@ export const deleteProductController = async (request, response) => {
 
 export const updateProductController = async (request, response) => {
   let id = request.params.pid;
-  let data = request.body;
-  if (Object.keys(data).length < 7)
-    return response.sendRequestError("All the fields are needed");
+  let product = request.body;
 
+  if (
+    !product.title ||
+    !product.description ||
+    !product.price ||
+    !product.thumbnail ||
+    !product.stock ||
+    !product.code
+  ) {
+    return handleError.createError({
+      name: "Product update error",
+      cause: generateProductError(product),
+      message: "Error trying to update the product",
+      code: EErrors.INVALID_PRODUCT_ERROR,
+    });
+  }
   try {
-    const result = await productRepository.updateProduct(id, data);
+    const result = await productRepository.updateProduct(id, product);
     if (result.error) {
       return response.sendRequestError(result.error);
     }
